Moved the PaginationConstants to a 'Constants' directory.
