+++ Servers/UI/OJS.Servers.Ui/Controllers/SubmissionsController.cs
@@ -18,7 +18,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using static Microsoft.AspNetCore.Http.StatusCodes;
 using static OJS.Common.GlobalConstants.Roles;
-using static OJS.Services.Common.PaginationConstants.Submissions;
+using static OJS.Services.Common.Constants.PaginationConstants.Submissions;
 
 public class SubmissionsController : BaseApiController
 {
