+++ Services/UI/OJS.Services.Ui.Data/Implementations/ContestsDataExtensions.cs
@@ -2,12 +2,11 @@ namespace OJS.Services.Ui.Data.Implementations;
 
 using OJS.Data.Models.Contests;
 using OJS.Services.Infrastructure.Extensions;
-using OJS.Common.Extensions;
 using OJS.Services.Infrastructure.Models;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using static OJS.Services.Common.PaginationConstants.Contests;
+using static OJS.Services.Common.Constants.PaginationConstants.Contests;
 
 public static class ContestsDataExtensions
 {
